requirements.txt
matplotlib==3.7.0
numpy==1.24.2
Pillow==10.0.0
tensorflow_macos==2.13.0
